Refactor file operations: move temporary file handling to a dedicated test utilities module
